V1.4.6 优化主页的逻辑；重构播放列表框架 [author]        wm [description] 解决主页播放按钮不能同步播放状态的问题； 修改主页给Fragment传递service的时机； 优化MainViewFragment的初始化流程； 简化Handle的调用流程； 重构播放列表框架：规划后续只显示[当前列表、收藏列表、最近播放列表]，为了方便管理，需要所有列表共用同一套框架，在Fragment和Apdater内部对不同的列表做判断处理

+++ MediaPlayProject/app/src/main/java/com/example/mediaplayproject/adapter/MusicListAdapter.java
@@ -21,20 +21,26 @@ import java.util.List;
 
 /**
  * @author wm
+ * @Classname MusicListAdapter
+ * @Description TODO
+ * @Version 1.0.0
+ * @Date 2023/8/26 14:57
+ * @Created by wm
  */
-public class FavoriteMusicAdapter extends BaseAdapter {
+public class MusicListAdapter extends BaseAdapter {
     private final Context mContext;
     private List<MediaFileBean> musicInfoList;
-
     private int defaultSelection = -1;
     private int text_selected_color;
     private int bg_selected_color;
     private ColorStateList colors;
+    private int listMode;
 
-
-    public FavoriteMusicAdapter(Context mContext, List<MediaFileBean> musicInfoList) {
+    @SuppressLint("UseCompatLoadingForColorStateLists")
+    public MusicListAdapter(Context mContext, int listMode, List<MediaFileBean> musicList) {
         this.mContext = mContext;
-        this.musicInfoList = musicInfoList;
+        this.listMode = listMode;
+        this.musicInfoList = musicList;
         Resources resources = mContext.getResources();
         // 文字选中的颜色
         text_selected_color = resources.getColor(R.color.text_pressed);
@@ -43,10 +49,6 @@ public class FavoriteMusicAdapter extends BaseAdapter {
         // 文字未选中状态的selector
         colors = mContext.getResources().getColorStateList(R.color.listview_text_color_selector);
         resources = null;
-    }
-
-    public FavoriteMusicAdapter(Context mContext) {
-        this.mContext = mContext;
     }
 
     @Override
@@ -69,55 +71,49 @@ public class FavoriteMusicAdapter extends BaseAdapter {
     public View getView(final int position, View convertView, ViewGroup parent) {
         ViewHolder holder;
         if (convertView == null) {
-            //获取LayoutInflater实例
-            convertView = LayoutInflater.from(mContext).inflate(R.layout.favorite_music_item, null);
+            convertView = LayoutInflater.from(mContext).inflate(R.layout.music_list_item, null);
             holder = new ViewHolder();
-            holder.tvFavoriteName = convertView.findViewById(R.id.tv_favorite_music_name);
-            holder.ivClose = convertView.findViewById(R.id.iv_delete_music);
+            holder.tvMusicName = convertView.findViewById(R.id.tv_music_name);
+            holder.ivDeleteMusic = convertView.findViewById(R.id.iv_delete_music);
             //将Holder存储到convertView中
             convertView.setTag(holder);
         } else {
-            //convertView不为空时，从convertView中取出Holder
+            // convertView不为空时，从convertView中取出Holder
             holder = (ViewHolder) convertView.getTag();
         }
-        holder.tvFavoriteName.setText(musicInfoList.get(position).getTitle());
-        holder.ivClose.setImageResource(R.mipmap.close);
+        holder.tvMusicName.setText(musicInfoList.get(position).getTitle());
+        holder.ivDeleteMusic.setImageResource(R.mipmap.ic_delete);
         if (position == defaultSelection) {
-            //设置已选择选项的颜色
-            holder.tvFavoriteName.setTextColor(text_selected_color);
-            //可以设置背景颜色，这里不设置
-            //convertView.setBackgroundColor(bg_selected_color);
+            // 设置已选择选项的颜色
+            holder.tvMusicName.setTextColor(text_selected_color);
         } else {
-            //设置按压效果
-            holder.tvFavoriteName.setTextColor(colors);
+            holder.tvMusicName.setTextColor(colors);
         }
-        //监听item里面的收藏按钮事件，需要在自定义Adapter的getView方法首个参数前添加final关键字(final int position...)
-        convertView.findViewById(R.id.iv_delete_music).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //取消收藏，这里调用deleteMusicFromFavoriteList删除收藏歌曲后，当前的收藏列表无需再删除，因为是同一对象
+        if (listMode == Constant.LIST_MODE_DEFAULT) {
+            // 默认列表不支持删除操作
+            holder.ivDeleteMusic.setEnabled(false);
+            holder.ivDeleteMusic.setVisibility(View.GONE);
+        } else {
+            // 监听item里面的删除按钮事件，需要在自定义Adapter的getView方法首个参数前添加final关键字(final int position...)
+            convertView.findViewById(R.id.iv_delete_music).setOnClickListener(v -> {
                 DataRefreshService.deleteMusicFromFavoriteList(musicInfoList.get(position));
                 DebugLog.debug("delete position " + position);
                 notifyDataSetChanged();
-            }
-        });
+            });
+        }
+
         return convertView;
+    }
+
+    static class ViewHolder {
+        TextView tvMusicName;
+        ImageView ivDeleteMusic;
     }
 
     public void notifyItemChanged(int position) {
         notifyDataSetChanged();
     }
 
-    static class ViewHolder {
-        TextView tvFavoriteName;
-        ImageView ivClose;
-    }
-
-    /**
-     * @param position -1：取消高亮效果； 0 < position < size(): 设置高亮效果
-     * @createTime 2023/2/11 15:37
-     * @description 设置高亮的下标
-     */
     public void setSelectPosition(int position) {
         if (position == -1) {
             //若传进来的值是-1，则代表要取消播放歌曲高亮效果
@@ -132,7 +128,7 @@ public class FavoriteMusicAdapter extends BaseAdapter {
 
     /**
      * @createTime 2023/2/11 14:56
-     * @description 专门用于--删除收藏歌曲的时候，若删除的下标小于当前播放的下标，列表会上移，
+     * @description 删除歌曲的时候，若删除的下标小于当前播放的下标，列表会上移，
      * 但是高亮的位置不变，导致了正在播放的高亮位置不正确，所以需要刷新指向高亮的下标
      */
     public void refreshSelectedPosition() {
@@ -144,7 +140,6 @@ public class FavoriteMusicAdapter extends BaseAdapter {
 
     /**
      *  检查要删除的歌曲下标与当前播放歌曲下标的比较情况
-     *  @author wm
      *  @createTime 2023/8/17 16:30
      * @param deletePosition: 要删除的歌曲下标
      * @return : int
